requirements.txt
schedule
python-dotenv
requests
beautifulsoup4
